+++ CHANGELOG.md
@@ -8,6 +8,16 @@
 
     *Joel Hawksley*
 
+* Allow copying from elements using `for` in `ClipboardCopy`.
+
+    *Manuel Puyol*
+
+### Breaking changes
+
+* Remove `label` argument in favor of `aria-label` in `ClipboardCopy`.
+
+    *Manuel Puyol*
+
 ### Misc
 
 * Add autocorrect for button linters.